Re-consider use of project_root
Even though project_root will have a constant value, there's no benefit
to capitalising it. My thinking hasn't settled on this yet, however
intuitively it makes sense to keep it lower case unless there's a more
compelling reason to capitalise it.

I suspect that the upper case version of constants ("don't change these
variables!") will end up closer to the code that uses them e.g. at the
top of classes, i.e. as class variables. Or at the top of functions. In
those instances, those variables (that shouldn't change) make more sense
as lower case. Unless we're explicitly drawing attention to the fact
they must not change. #tobediscussedfurther

We've also added a "config.py" module to expose global variables like
"project_root" to the rest of the app and not just in the tests.