Expand validation scripts into 'codetidy' package
We're expanding this project to manage all kinds of code tidy up actions
like: validate doctrings, validate return types, find poor names etc.

The name 'codetidy' seems like a suitable package name for this group of
activities.

--

In more detail

We were getting some duplication across scripts so it is time to
refactor. We extracted `CodeBlock` into its own class so we could get
access to the function definitions, nodes, tree etc for a specific code
block.

We also use the strategy pattern to combine filters together and filter
the function definitions by those filters. For instance we want to find
all docstrings that are adjacent to the code. Or find any standalone
functions or methods that are missing return types.

Note: The `ast.FunctionDef` discovers both standalone functions and
methods inside classes. In the ast module, `ast.FunctionDef` is used to
represent both functions and methods. In Python's abstract syntax tree
(AST), there is no separate node type for methods; both functions and
methods are represented by ast.FunctionDef nodes.

We've also added a script to find poor names in the code base so we can
tidy those up.

Now we have three types of validation:
- Find poor names
- Validate docstrings
- Validate return types

The rest can be done via Pylance and the pyrightconfig.json config file.